Attempt to parse the exception string from environment provider
If the environment provider returns an exception the ESR would
raise a generic Exception and, in the case of Checkout failures
we sometimes would raise Exception('') which gives us nothing.
Added a new method that parses the last part of a traceback to
try to find the actual exception raised and re-raise it.